requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
django-multiuploader==0.2.9
djangorestframework==3.12.2
Pillow==8.1.1
psycopg2-binary==2.8.6
python-decouple==3.4
python-magic==0.4.22
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
